# docs/swagger.yaml
definitions:
...
  domain.CRMCustomer:
    properties:
      id:
        type: string
      title:
        type: string
    type: object
...
  domain.Layout:
    properties:
      id:
        type: string
      is_active:
        type: boolean
      kind:
        type: string
      owner:
        $ref: '#/definitions/domain.CRMCustomer'
        type: object
      title:
        type: string
    type: object
...
  domain.Layouts:
    items:
      $ref: '#/definitions/domain.Layout'
    type: array
...
  domain.MallEntrance:
    properties:
      created_at:
        type: string
      creator:
        type: string
      entrance_id:
        type: string
      floor_id:
        type: string
      is_active:
        type: boolean
      kind:
        type: string
      layout_id:
        type: string
      modified_at:
        type: string
      modifier:
        type: string
      notes:
        type: string
      options:
        type: string
      sensors:
        $ref: '#/definitions/domain.MallSensors'
        type: object
      title:
        type: string
      valid_from:
        type: string
      valid_to:
        type: string
    type: object
  domain.MallEntrances:
    items:
      $ref: '#/definitions/domain.MallEntrance'
    type: array
  domain.MallSensor:
    properties:
      created_at:
        type: string
      creator:
        type: string
      device_id:
        type: string
      external_id:
        type: string
      is_active:
        type: boolean
      kind:
        type: string
      layout_id:
        type: string
      modified_at:
        type: string
      modifier:
        type: string
      notes:
        type: string
      options:
        type: string
      sensor_id:
        type: string
      title:
        type: string
      valid_from:
        type: string
      valid_to:
        type: string
    type: object
  domain.MallSensors:
    items:
      $ref: '#/definitions/domain.MallSensor'
    type: array
...
  infra.ErrResponse:
    properties:
      code:
        description: application-specific error code
        type: integer
      error:
        description: application-level error message, for debugging
        type: string
      status:
        description: user-level status message
        type: string
    type: object
  infra.HTTPError:
    properties:
      message:
        type: object
    type: object
  infra.LayoutResponse:
    properties:
      data:
        $ref: '#/definitions/domain.Layouts'
        type: object
      result_set:
        $ref: '#/definitions/infra.ResultSet'
        type: object
    type: object
...
  infra.MallEntrancesResponse:
    properties:
      data:
        $ref: '#/definitions/domain.MallEntrances'
        type: object
      result_set:
        $ref: '#/definitions/infra.ResultSet'
        type: object
    type: object
...
  infra.ResultSet:
    properties:
      count:
        type: integer
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
info:
  contact:
    email: 1020@watcom.ru
    name: API Support
    url: https://helpdesk.watcom.ru
  description: |-
    This is a general service for interacting with layout configuration.
    Общий сервис для взаимодействия с конфигурацией схем размещения.
  license: {}
  title: LayoutConfig API
  version: 2.0/1.1.21
paths:
...
  /v2/layouts:
    get:
      description: get slice of layouts/projects configuration with location, date,
        offset, limit parameters
      parameters:
      - description: default=ru
        in: query
        name: loc
        type: string
      - description: YYYY-MM-DD, default=today
        in: query
        name: date
        type: string
      - description: default=0
        in: query
        name: offset
        type: integer
      - description: default=20
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      - text/xml
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/infra.LayoutResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/infra.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/infra.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/infra.ErrResponse'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/infra.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/infra.ErrResponse'
      summary: Get all layouts
      tags:
      - common
  /v2/layouts/{layout_id}:
    get:
      description: get layout with location, date and layout_id parameters
      parameters:
      - description: digit/uuid format
        in: path
        name: layout_id
        required: true
        type: string
      - description: default=ru
        in: query
        name: loc
        type: string
      - description: YYYY-MM-DD, default=today
        in: query
        name: date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Layout'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/infra.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/infra.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/infra.ErrResponse'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/infra.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/infra.ErrResponse'
      summary: Get specified layout
      tags:
      - common
...
  /v2/malls/entrances:
    get:
      description: |-
        get slice of entrances with loc, date, layout_id, floor_id, offset, limit, fields parameters
        layout_id (recommended parameter), if not pass datasource may be not correct
        fields - comma separated values of field names, can be entrance_id,layout_id,floor_id,kind,title...
      parameters:
      - description: uuid/digit format, default=*
        in: query
        name: layout_id
        type: string
      - description: default=*
        in: query
        name: floor_id
        type: string
      - description: default=entrance
        in: query
        name: kind
        type: string
      - description: location, default=ru
        in: query
        name: loc
        type: string
      - description: ISO8601 YYYY-MM-DD date
        in: query
        name: date
        type: string
      - description: comma separated list ids, default=*
        in: query
        name: entrance_ids
        type: string
      - description: default=0
        in: query
        name: offset
        type: integer
      - description: default=20
        in: query
        name: limit
        type: integer
      - description: entrance_id,floor_id,title... default=all
        in: query
        name: fields
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/infra.MallEntrancesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/infra.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/infra.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/infra.ErrResponse'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/infra.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/infra.ErrResponse'
      summary: Get all entrances in the mall schema
      tags:
      - malls/entrances
  /v2/malls/entrances/{entrance_id}:
    get:
      description: |-
        get specified entrance with specified, loc, date, fields parameters
        layout_id (recommended parameter), if not pass datasource may be not correct
        fields - comma separated values of field names, can be entrance_id,layout_id,floor_id,kind,title...
      parameters:
      - description: uuid/digit format, default=*
        in: query
        name: layout_id
        type: string
      - description: uuid format
        in: path
        name: entrance_id
        required: true
        type: string
      - description: location, default=ru
        in: query
        name: loc
        type: string
      - description: ISO8601 YYYY-MM-DD date
        in: query
        name: date
        type: string
      - description: entrance_id,floor_id,title... default=all
        in: query
        name: fields
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.MallEntrance'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/infra.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/infra.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/infra.ErrResponse'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/infra.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/infra.ErrResponse'
      summary: Get specified entrance in the retail schema
      tags:
      - malls/entrances
...
swagger: "2.0"